+++ functions/__init__.py
@@ -1,4 +1,4 @@
 from .twitter import Twitter
-from .danbooru import Danbooru
+from .scrape_websites.danbooru import Danbooru
 from .time_log import Time_Log
 from .discord_stdout import Discord_Stdout
