+++ functions/scrape_websites/__init__.py
@@ -1,0 +1,2 @@
+from .scrape import Scrape
+from .danbooru import Danbooru
